# 3016 3107 performance optimizations