--- a/app/controllers/gobierto_budgets/application_controller.rb
+++ b/app/controllers/gobierto_budgets/application_controller.rb
@@ -13,4 +13,10 @@ class GobiertoBudgets::ApplicationController < ApplicationController
   def set_current_site
     @site = SiteDecorator.new(current_site)
   end
+
+  private
+
+  def cache_path
+    "#{current_site.cache_key}/#{current_module}/#{self.controller_name}/#{self.action_name}/#{I18n.locale}"
+  end
 end