--- a/app/controllers/gobierto_budgets/application_controller.rb
+++ b/app/controllers/gobierto_budgets/application_controller.rb
@@ -1,6 +1,8 @@
 class GobiertoBudgets::ApplicationController < ApplicationController
   include User::SessionHelper
 
+  helper_method :cache_path
+
   rescue_from GobiertoBudgets::BudgetLine::RecordNotFound, with: :render_404
   rescue_from GobiertoBudgets::BudgetLine::InvalidSearchConditions do |exception|
     head :bad_request