--- a/app/controllers/meta_welcome_controller.rb
+++ b/app/controllers/meta_welcome_controller.rb
@@ -30,7 +30,7 @@ class MetaWelcomeController < ApplicationController
 
       @page = GobiertoCms::PageDecorator.new(page)
 
-      render "gobierto_cms/pages/show", layout: GobiertoCms::ApplicationController::DEFAULT_LAYOUT
+      render "gobierto_cms/pages/meta_welcome", layout: GobiertoCms::ApplicationController::DEFAULT_LAYOUT
     end
   end
 end