--- a/app/models/concerns/gobierto_common/sectionable.rb
+++ b/app/models/concerns/gobierto_common/sectionable.rb
@@ -5,30 +5,27 @@ module GobiertoCommon
     extend ActiveSupport::Concern
 
     included do
-      def section_id
-        return if GobiertoCms::SectionItem.where(item: self).joins(:section).empty?
+      has_many :section_items, as: :item, class_name: "GobiertoCms::SectionItem"
+      has_many :sections, through: :section_items
 
-        GobiertoCms::SectionItem.where(item: self).first.section.id
+      def section
+        sections.first
+      end
+
+      def section_id
+        section&.id
       end
 
       def parent_id
-        unless GobiertoCms::SectionItem.where(item: self).empty?
-          if GobiertoCms::SectionItem.where(item: self).first.parent_id == 0
-            0
-          else
-            GobiertoCms::SectionItem.where(item: self).first.parent.try(:id)
-          end
-        else
-          nil
-        end
+        return unless section_items.exists?
+
+        section_item = section_items.first
+
+        section_item.parent_id.zero? ? 0 : section_item.parent&.id
       end
 
       def position
-        unless GobiertoCms::SectionItem.where(item: self).empty?
-          GobiertoCms::SectionItem.where(item: self).first.position
-        else
-          nil
-        end
+        section_items.first&.position
       end
     end
   end