--- a/app/views/gobierto_cms/shared/_page.html.erb
+++ b/app/views/gobierto_cms/shared/_page.html.erb
@@ -1,0 +1,1 @@
+<li><%= link_to page.title, gobierto_cms_page_or_news_path(page) %></li>