--- a/app/views/gobierto_cms/shared/_pages_side_navigation.html.erb
+++ b/app/views/gobierto_cms/shared/_pages_side_navigation.html.erb
@@ -2,13 +2,12 @@
   <% if @section %>
     <!-- This is not clear. Can't we generate the tree just passing the section instead of two arguments? -->
     <%= section_tree(@section.section_items.first_level.not_archived.not_drafted, @section_item.hierarchy_and_children(only_public: true)) %>
+
   <% elsif @collection %>
     <ul>
       <li><%= link_to @collection.title, gobierto_cms_pages_path(@collection.slug) %></a>
         <ul>
-          <% @pages.each do |page| %>
-            <li><%= link_to page.title, gobierto_cms_page_or_news_path(page) %></li>
-          <% end %>
+          <%= render partial: "gobierto_cms/shared/page", collection: @pages, cached: true %>
         </ul>
       </li>
     </ul>