… app/views/gobierto_budgets/budget_lines/show.html.erb