Closes #3107
Closes #3106

## :v: What does this PR do?

* Caches a fragment of meta_welcome view which loads slowly
* Refactors some code and uses eager load of some associations to avoid unnecessary queries
* Uses its own view for the meta_welcome page and caches the content depending on locale and site
* Makes CMS resources belonging to site to touch it on updates
* Improves performance of application in gobierto budgets home page and budget lines show and index actions:
  * Using caches_action from controller for not signed in users
  * Using fragment caching from view in show partial

## :mag: How should this be manually tested?

For meta_welcome:
* An index with a lot of subpages was taking around 600ms - 1s
* The current time is about 200ms

For gobierto_budgets:
* Before this changes, the time to load a budget line in staging, taken from browser info has an average of 1.63s with maximum of 3.40s and minimum of 1.06s, regardless the user is logged in or not
* After the changes, in staging:
  * For not signed in users: Average: 0.78, max: 0.87, min: 0.71
  * For signed in users: Average: 0.92, max: 0.99, min: 0.82

Looking at appsignal the performance seems to be much better than these times: https://appsignal.com/populate/sites/5eb0fcfc74782044d1198606/performance/incidents/24?timerange=Wed+May+27+2020+20%3A39%3A00+GMT%2B0200%2CWed+May+27+2020+20%3A40%3A00+GMT%2B0200&incident_tab=host,  from orders of 10^3 to 10 milliseconds to but I'm not sure if I'm reading the stats correctly

## :shipit: Does this PR changes any configuration file?

No

(Changes in these files might need to update the role in Ansible)

## :book: Does this PR require updating the documentation?

No